+++ CHANGELOG.md
@@ -11,3 +11,4 @@
  - pull all `dailySummaries` for a user. The old data is deleted and it is reinserted
  - pull all `activities` for a user. The old data is deleted and it is reinserted
  - save `fitrockr_activities_summary` for a user. The old data is deleted and it is reinserted
+ - add manual pull for a selected user